Better organize types
